pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcelo Coelho <92119579+MCoelho222@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"
google-auth = "^2.14.1"
google-auth-oauthlib = "^0.7.1"
pymongo = "^4.3.3"
Flask-PyMongo = "^2.3.0"
Flask-Cors = "^3.0.10"
PyJWT = "^2.6.0"
google-api-python-client = "^2.66.0"
google-auth-httplib2 = "^0.1.0"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
